pyproject.toml
[project]
name = "evals-workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipywidgets>=8.1.7",
    "marimo>=0.16.5",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "pydantic>=2.12.2",
    "duckdb>=1.1.2",
    "pydantic-ai>=0.0.15",
    "ruff>=0.14.0",
    "python-dotenv>=1.1.1",
    "flask>=3.1.2",
    "scikit-learn>=1.7.2",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []
